From Humar Carmona, "When reading a DXF file, the reader breaks at a debug assertion at vector (it breaks on release version). Inspecting the code show that the cause could be in dxfEntity.cpp.
It seems that the problem is an offending "short" used in a for loop, where it should be "int" or "long". It causes an index out of range error. "

Merged from svn trunk using:
svn merge -r 9803:9804 http://www.openscenegraph.org/svn/osg/OpenSceneGraph/trunk/src/osgPlugins/dxf/dxfEntity.cpp

+++ src/osgPlugins/dxf/dxfEntity.cpp
@@ -597,7 +597,7 @@ dxfPolyline::drawScene(scene* sc)
         // is there a flag 1 or 32 for 64?
     } else {
         // simple polyline?
-        for (short i = _vertices.size()-1; i >= 0; i--)
+        for (int i = _vertices.size()-1; i >= 0; i--)
             vlist.push_back(_vertices[i]->getVertex());
         if (_flag & 1) {
 //            std::cout << "line loop " << _vertices.size() << std::endl;
